task 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -32,8 +32,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if (!user) {
+    return response.status(400).json({ error: 'UserId missing or not valid' })
+  } else {
+    const blog = await Blog.findById(request.params.id)
+    if (!blog){
+      return response.status(400).json({ error: 'BlogId is not valid' })
+    } else if ( blog.user.toString() === user.id.toString() ){
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(400).json({ error: 'UserId associated with blog is not matching with current user' })
+    }
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
